admin 세션 넣어봤는데 등록에
admin 세션 넣어봤는데 등록에

+++ src/main/java/com/busanbank/card/admin/dto/AdminDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class AdminDto {
 
-	private int admin_no;
+	private Long admin_no;
 	private String username;
 	private String password;
 	private String name;
